cleaned manifest + removed jc full screen ads
