+++ AppLovin MAX Demo App - Kotlin/app/src/main/java/com/applovin/enterprise/apps/demoapp/ads/applovin/banners/BannerJetpackComposeActivity.kt
@@ -47,7 +47,8 @@ class BannerJetpackComposeActivity : BaseJetpackComposeAdActivity() {
                     TextButton(
                         onClick = { adLoader.loadAd() },
                         modifier = Modifier.align(Alignment.CenterHorizontally)
-                    ) {
+                    )
+                    {
                         Text("Load")
                     }
                     AppLovinAdComposable(adLoader)
