+++ AppLovin MAX Demo App - Kotlin/app/src/main/java/com/applovin/enterprise/apps/demoapp/ads/applovin/rewarded/RewardedVideosDemoMenuActivity.kt
@@ -8,7 +8,6 @@ class RewardedVideosDemoMenuActivity : DemoMenuActivity()
 {
     override fun getListViewContents(): Array<DemoMenuItem> = arrayOf(
             DemoMenuItem("Basic Integration", Intent(this, RewardedVideosActivity::class.java)),
-            DemoMenuItem("Zone Integration", Intent(this, RewardedVideosZoneActivity::class.java)),
-            DemoMenuItem("Jetpack Compose Integration", Intent(this, RewardedVideosJetpackComposeActivity::class.java))
+            DemoMenuItem("Zone Integration", Intent(this, RewardedVideosZoneActivity::class.java))
     )
 }
