+++ AppLovin MAX Demo App - Kotlin/app/src/main/res/layout/activity_interstitial_ad.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ads.max.interstitial.BasicIntegrationInterstitialAdActivity">
+    tools:context=".ads.max.interstitial.InterstitialAdActivity">
 
     <androidx.recyclerview.widget.RecyclerView
         android:id="@+id/callbacksRecyclerView"
