+++ AppLovin MAX Demo App - Kotlin/app/src/main/res/layout/activity_rewarded_ad.xml
@@ -4,7 +4,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     xmlns:app="http://schemas.android.com/apk/res-auto"
-    tools:context=".ads.max.rewarded.BasicIntegrationRewardedAdActivity">
+    tools:context=".ads.max.rewarded.RewardedAdActivity">
 
     <androidx.recyclerview.widget.RecyclerView
         android:id="@+id/callbacksRecyclerView"
