+++ AppLovin MAX Demo App - Kotlin/app/src/main/res/layout/activity_rewarded_videos.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ads.max.rewarded.BasicIntegrationRewardedAdActivity">
+    tools:context=".ads.max.rewarded.RewardedAdActivity">
 
     <Button
         android:id="@+id/loadButton"
