+++ AppLovin MAX Demo App - Kotlin/build.gradle
@@ -19,7 +19,7 @@ allprojects {
     repositories {
         google()
         jcenter()
-        
+
     }
 }
